feat(mtpndd): add multi-terminal node support with fraction and double leaves
Extend MTPNDD to represent functions to arbitrary rationals (and IEEE
doubles) instead of just Boolean indicators.  Terminal nodes now share
the existing node struct via a tagged-by-field_id leaf union, with two
canonical leaf tables backing the type-specific dedup.

Data layer
- mtpndd_node_t shrinks from 24 to 16 bytes: ref_count is now uint32,
  and `edges`/`leaf_value` share a union.  `field_id` in the upper
  range (>= MTPNDD_LEAF_FIELD_ID_MIN) marks a leaf, with the specific
  value selecting fraction (0xFFFFFFFF) vs double (0xFFFFFFFE).
- MTPNDD_TRUE/FALSE keep their addresses but are seeded into the
  fraction leaf table as 1/1 and 0/1 sentinels, so make_fraction
  canonicalizes 0 -> &MTPNDD_FALSE and 1 -> &MTPNDD_TRUE.
- New mtpndd_leaf_table module: sharded-spinlock canonical table keyed
  on packed leaf_value, used by both fraction and double tables.

Leaf API (mtpndd_leaf.{h,c})
- mtpndd_make_fraction(int32, int32) with int64 gcd normalization.
  denom must be non-zero; sign moves to numerator; values that exceed
  int32 after gcd return INVALID_PARAM.
- mtpndd_make_double(double) keyed on the IEEE bit pattern.
- mtpndd_is_leaf / is_fraction_leaf / is_double_leaf accessors.

Arithmetic (mtpndd_arith.{h,c})
- plus / minus / times / divide dispatch on leaf type at the leaf-leaf
  case; mixing fraction and double is an error.  Fraction arithmetic
  uses int64 intermediates and re-normalizes via int64 gcd.
- plus and times reuse the existing or_cache and and_cache (Boolean
  and MT modes are mutually exclusive in practice); minus and divide
  are uncached in phase 1.
- Same-field ops use AND-with-NOT residuals like mtpndd_or_rec for
  plus/minus; minus's B-side residuals go through a recursive negate.
- divide checks that b's support covers a's; it errors on division by
  zero anywhere in the support.
- plus and times are parallelized via Lace SPAWN sub-tasks
  (mtpndd_plus_same_pair / plus_diff_pair / times_*); reuses
  MTPNDD_SPAWN_THRESHOLD heuristic and a 256-pair flush threshold.
  minus / divide / abstract_plus stay sequential.
- abstract_plus(node, field) sums out one field assuming labels at
  that field only depend on its own BDD vars.  Three cases:
  field_id == target collapses (sum of child * |label|);
  field_id < target recurses; field_id > target scales by 2^bits.
- leafcount() walks the DAG with a dedup hash set.

satcount extension
- mtpndd_satcount_rec treats any non-zero terminal (fraction with
  numer != 0, or double with v != 0.0) as contributing the full
  measure; zero leaves contribute 0.  Boolean semantics preserved.

Tests / benchmark
- test/fraction.c: 10 sub-tests covering canonicalization, leaf
  arithmetic, int32 overflow detection, double leaves, mixed-type
  rejection, internal plus, minus residual via negate, times
  disjoint, divide support check, abstract_plus, parallel plus.
- test/fraction_benchmark.c: 2-field dense MTPNDD plus/times benchmark
  parameterized by worker count and log2(cubes).  At 64 cubes shows
  ~1.6x speedup on 4 workers; larger workloads regress and are noted
  as future tuning work.

+++ sylvan/src/sylvan/mtpndd/mtpndd_nodetable.c
@@ -248,8 +248,8 @@ mtpndd_error_t mtpndd_ref(mtpndd_t *node) {
         mtpndd_set_error(MTPNDD_ERROR_NULL_POINTER, __func__, __LINE__);
         return MTPNDD_ERROR_NULL_POINTER;
     }
-    if (node->ref_count == UINT64_MAX) {
-        // Terminal nodes
+    if (node->ref_count == UINT32_MAX) {
+        // Terminal / protected nodes
         return MTPNDD_SUCCESS;
     }
     atomic_fetch_add(&node->ref_count, 1);
@@ -261,8 +261,8 @@ mtpndd_error_t mtpndd_deref(mtpndd_t *node) {
         mtpndd_set_error(MTPNDD_ERROR_NULL_POINTER, __func__, __LINE__);
         return MTPNDD_ERROR_NULL_POINTER;
     }
-    if (node->ref_count == UINT64_MAX) {
-        // Terminal nodes
+    if (node->ref_count == UINT32_MAX) {
+        // Terminal / protected nodes
         return MTPNDD_SUCCESS;
     }
     atomic_fetch_sub(&node->ref_count, 1);
@@ -274,7 +274,7 @@ mtpndd_error_t mtpndd_protect(mtpndd_t *node) {
         mtpndd_set_error(MTPNDD_ERROR_NULL_POINTER, __func__, __LINE__);
         return MTPNDD_ERROR_NULL_POINTER;
     }
-    atomic_init(&node->ref_count, UINT64_MAX);
+    atomic_init(&node->ref_count, UINT32_MAX);
     return MTPNDD_SUCCESS;
 }
 
@@ -658,7 +658,7 @@ static size_t mtpndd_gc_sweep(void) {
             while (entry) {
                 mtpndd_nodetable_bucket_entry_t *next_entry = entry->next;
                 mtpndd_node_t *node = entry->node;
-                uint64_t refc = atomic_load_explicit(&node->ref_count, memory_order_relaxed);
+                uint32_t refc = atomic_load_explicit(&node->ref_count, memory_order_relaxed);
                 if (refc == 0) {
                     mtpndd_release_node(nodetable, i, entry, node);
                     reclaimed++;
